# openapi.yaml
openapi: 3.1.0
info:
  title: Traccar
  description: Traccar GPS tracking server API documentation. To use the API you need to have a server instance. For testing purposes you can use one of free [demo servers](https://www.traccar.org/demo-server/). For production use you can install your own server or get a [subscription service](https://www.traccar.org/product/tracking-server/).
  contact:
    name: Traccar Support
    url: https://www.traccar.org/
    email: support@traccar.org
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: '6.15.3'
servers:
  - url: https://demo.traccar.org/api
    description: Demo Server 1
  - url: https://demo2.traccar.org/api
    description: Demo Server 2
  - url: https://demo3.traccar.org/api
    description: Demo Server 3
  - url: https://demo4.traccar.org/api
    description: Demo Server 4
  - url: https://server.traccar.org/api
    description: Subscription Server
  - url: http://{host}:{port}/api
    description: Other Server
    variables:
      host:
        default: localhost
      port:
        enum:
          - '8082'
          - '80'
        default: '8082'
security:
  - BasicAuth: []
  - ApiKey: []
...
paths:
...
  /drawings:
    get:
      operationId: getDrawings
      summary: Fetch map drawings visible to the current user
      tags:
        - Drawings
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Drawing'
        '401':
          description: Unauthorized
          content: {}
    post:
      operationId: postDrawings
      summary: Create a map drawing
      tags:
        - Drawings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Drawing'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drawing'
        '400':
          description: Invalid drawing payload
          content: {}
        '401':
          description: Unauthorized
          content: {}
  /drawings/{id}:
    get:
      operationId: getDrawingsId
      summary: Fetch a map drawing visible to the current user
      tags:
        - Drawings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drawing'
        '404':
          description: Not found
          content: {}
    put:
      operationId: putDrawingsId
      summary: Update an owned map drawing
      tags:
        - Drawings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Drawing'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drawing'
        '400':
          description: Invalid drawing payload
          content: {}
        '404':
          description: Not found
          content: {}
    delete:
      operationId: deleteDrawingsId
      summary: Delete an owned map drawing, or any drawing as an administrator
      tags:
        - Drawings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted
          content: {}
        '404':
          description: Not found
          content: {}
...
components:
  schemas:
...
    Drawing:
      type: object
      required:
        - type
        - geometry
        - color
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          description: Unique drawing identifier
        ownerId:
          type: integer
          format: int64
          readOnly: true
          description: Identifier of the user that owns the drawing
        ownerName:
          type: string
          readOnly: true
          description: Current display name of the drawing owner
        type:
          type: string
          enum: [line, arrow, polygon, rectangle, circle, text]
        geometry:
          type: object
          description: WGS84 GeoJSON geometry appropriate for the drawing type
          required:
            - type
            - coordinates
          properties:
            type:
              type: string
              enum: [Point, LineString, Polygon]
            coordinates:
              type: array
              items: {}
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        text:
          type: [string, 'null']
          maxLength: 200
          description: Multiline label used by text drawings
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKey:
      type: http
      scheme: bearer